requirements.txt
# Core Data Handling
pandas==2.3.0
numpy==2.3.0

# Web Scraping
newspaper3k==0.2.8
googlesearch-python==1.2.3
beautifulsoup4==4.13.4
lxml==5.4.0
tldextract==5.3.0

# Sentiment Analysis
textblob==0.18.0
nltk==3.9.1

# Machine Learning & Forecasting
tensorflow==2.16.1
scikit-learn==1.4.2

# Interpretability
shap==0.45.0

# Visualization
matplotlib==3.8.4

# UI Dashboard
streamlit==1.35.0

# Utility (optional but used)
fpdf==1.7.2
tqdm==4.67.1
requests==2.32.4
python-dateutil==2.9.0.post0
